requirements.txt
utils-nuuuwan
wordcloud
matplotlib
numpy 
Pillow